+++ src/components/pages/pagebanners/TopBannerImg.vue
@@ -27,8 +27,8 @@
               accept="image/*"
               @input="uploadImg"
             />
-            <label class="custom-file-label" for="customFile"
-              >Выберите файл</label
+            <label class="custom-file-label text-center" for="customFile"
+              >Добавить</label
             >
           </div>
         </div>
@@ -99,9 +99,15 @@ export default {
   name: "TopBannerImg",
   methods: {
     removeTopBannerImg: function () {
-      const storageRef = firebase.storage();
-      let desertRef = storageRef.refFromURL(this.updatedImage);
-      desertRef.delete();
+      if (
+        !this.updatedImage.includes("blob") &&
+        !this.updatedImage.includes("noimage")
+      ) {
+        const storageRef = firebase.storage();
+
+        let desertRef = storageRef.refFromURL(this.updatedImage);
+        desertRef.delete();
+      }
 
       this.$emit("remove");
     },
@@ -116,38 +122,16 @@ export default {
 
       this.updatedImage = URL.createObjectURL(file);
 
-      // this.updatedImage = fetch(this.updatedImage)
-      //   .then((res) => {
-      //     return res.blob();
-      //   })
-      //   .then((blob) => {
-      //     //uploading blob to firebase storage
-      //     firebase
-      //       .storage()
-      //       .ref(`images/${Date.now()}${file.name}`)
-      //       .put(blob)
-      //       .then(function (snapshot) {
-      //         return snapshot.ref.getDownloadURL();
-      //       })
-      //       .then((url) => {
-      //         console.log("Firebase storage image uploaded : ", url);
-      //       });
-      //   })
-      //   .catch((error) => {
-      //     console.error(error);
-      //   });
-
-      const storageRef = firebase
-        .storage()
-        .ref(`images/${Date.now()}${file.name}`);
-      this.updatedImage = await storageRef
-        .put(file)
-        .then(async function (snapshot) {
-          return await snapshot.ref.getDownloadURL();
-        });
-
       this.$emit("update:image", this.updatedImage);
     },
   },
 };
 </script>
+
+<style scoped>
+.custom-file-label::after {
+  content: "";
+  background-color: transparent;
+  border-left: none;
+}
+</style>
