+++ src/components/pages/pagefilms/FilmCard.vue
@@ -11,13 +11,16 @@
     </div>
     <!-- /.card-header -->
     <div class="card-body d-flex flex-column" style="padding: 0">
-      <img
-        :src="updatedImage"
-        style="width: 100%; height: 100%"
-        class="mb-2"
-        alt=""
-      />
-      <span class="pb-2" style="text-align: center">{{ movieTitle }}</span>
+      <router-link to="/addfilm">
+        <img
+          :src="updatedImage"
+          style="width: 100%; height: 100%"
+          class="mb-2"
+          alt=""
+        />
+      </router-link>
+
+      <span class="pb-2 text-center">{{ movieTitle }}</span>
     </div>
     <!-- /.card-body -->
   </div>
@@ -41,9 +44,15 @@ export default {
   name: "FilmCard",
   methods: {
     removeFilmCard: function () {
-      const storageRef = firebase.storage();
-      let desertRef = storageRef.refFromURL(this.updatedImage);
-      desertRef.delete();
+      if (
+        !this.updatedImage.includes("blob") &&
+        !this.updatedImage.includes("noimage")
+      ) {
+        const storageRef = firebase.storage();
+
+        let desertRef = storageRef.refFromURL(this.updatedImage);
+        desertRef.delete();
+      }
 
       this.$emit("remove");
     },
